requirements.txt
django
dj-database-url
dj-static
gunicorn
psycopg2
static
django_jasmine
django_nose
South
django_compressor
BeautifulSoup
djangorestframework==2.3.12
markdown
pyyaml
django-userena
mock
django-simple-captcha
mandrill==1.0.6
django-mandrill==0.1.0